# public/swagger.yaml
components:
...
    schemas:
        TokenResponse:
            properties:
                id:
                    type: number
                    format: double
                    description: 'Access Token and Refresh Tokens'
                    example: someRandomCryptoString
                name:
                    type: string
                isActivated:
                    type: boolean
                tokens:
                    properties:
                        activationToken:
                            type: string
                        passwordResetToken:
                            type: string
                        refreshToken:
                            type: string
                        accessToken:
                            type: string
                    required:
                        - activationToken
                        - passwordResetToken
                        - refreshToken
                        - accessToken
                    type: object
            required:
                - id
                - name
                - isActivated
                - tokens
            type: object
            additionalProperties: false
        UserPayload:
            properties:
                name:
                    type: string
                    description: 'name for user'
                    example: 'John Snow'
                email:
                    type: string
                    description: 'email for user'
                    example: johnSnow01@gmail.com
                password:
                    type: string
                    description: 'Password for user'
                address:
                    type: string
                    description: 'address for user'
                    example: 'H#123 Block 2 Sector J, Abc Town, NY'
            required:
                - name
                - email
                - password
                - address
            type: object
            additionalProperties: false
        UserDetailsResponse:
            properties:
                email:
                    type: string
                name:
                    type: string
                address:
                    type: string
                profilePicture:
                    type: string
            required:
                - email
                - name
                - address
                - profilePicture
            type: object
            additionalProperties: false
        UserSearchResponse:
            properties:
                users:
                    items:
                        $ref: '#/components/schemas/UserDetailsResponse'
                    type: array
            required:
                - users
            type: object
            additionalProperties: false
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
info:
    title: express-ts
    version: 1.0.0
    license:
        name: ISC
    contact: {}
openapi: 3.0.0
paths:
    /auth/signup:
        post:
            operationId: SignUp
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {$ref: '#/components/schemas/TokenResponse'}
                                    - {type: string}
                            examples:
                                'Example 1':
                                    value: {name: 'John Snow', email: johnSnow01@gmail.com, address: 'H#123 Block 2 Sector J, Abc Town, NY', profilePicture: string}
            summary: 'Create user with the following attributes: email, Name, Password, address.'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserPayload'
    /auth/login:
        post:
            operationId: Login
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {$ref: '#/components/schemas/TokenResponse'}
                                    - {type: string}
                            examples:
                                'Example 1':
                                    value: {id: 0, name: Johny, isActivated: false, tokens: {accessToken: someRandomCryptoString, refreshToken: someRandomCryptoString, passwordResetToken: someRandomCryptoString, activationToken: someRandomCryptoString}}
            summary: 'logs user in and returns access token'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                password:
                                    type: string
                                email:
                                    type: string
                            required:
                                - password
                                - email
                            type: object
    /auth/logout:
        get:
            operationId: Logout
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: string
            summary: 'Verify and Removes JWT tokens and returns success message'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters: []
    /auth/activate:
        post:
            operationId: ActivateUser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {$ref: '#/components/schemas/TokenResponse'}
                                    - {type: string}
            summary: 'Verify and Removes JWT activationToken and returns success message'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                token:
                                    type: string
                                id:
                                    type: string
                            required:
                                - token
                                - id
                            type: object
    /auth/forgotpassword:
        post:
            operationId: ForgotPassword
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {$ref: '#/components/schemas/TokenResponse'}
                                    - {type: string}
                            examples:
                                'Example 1':
                                    value: {id: 0, name: Johny, isActivated: false, tokens: {accessToken: someRandomCryptoString, refreshToken: someRandomCryptoString, passwordResetToken: someRandomCryptoString, activationToken: someRandomCryptoString}}
            summary: 'sends a mail to user for password reset'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                email:
                                    type: string
                            required:
                                - email
                            type: object
    /auth/resendpasswordtoken:
        post:
            operationId: ResendPasswordToken
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {$ref: '#/components/schemas/TokenResponse'}
                                    - {type: string}
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                email:
                                    type: string
                            required:
                                - email
                            type: object
    /auth/resetpassword:
        post:
            operationId: ResetPassword
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: string
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                token:
                                    type: string
                                id:
                                    type: string
                                password:
                                    type: string
                            required:
                                - token
                                - id
                                - password
                            type: object
    '/user/{userId}':
        get:
            operationId: GetUser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDetailsResponse'
                            examples:
                                'Example 1':
                                    value: {name: 'John Snow', email: johnSnow01@gmail.com, address: 'H#123 Block 2 Sector J, Abc Town, NY', profilePicture: string}
            summary: 'Get user properties - such as name, email and address.'
            tags:
                - User
            security: []
            parameters:
                -
                    in: path
                    name: userId
                    required: true
                    schema:
                        type: string
    /user/search:
        post:
            operationId: SearchUser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserSearchResponse'
            tags:
                - User
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                query:
                                    type: string
                            required:
                                - query
                            type: object
    /user/updateprofilepic:
        put:
            operationId: UpdateProfilePic
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: string
            tags:
                - User
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                id:
                                    type: number
                                    format: double
                            required:
                                - id
                            type: object
...
servers:
    -
        url: /
...